move files to example dir to exclude them from coverage
